Rename stanza.resource to stanza.resources; rename resources.py to common.py
