+++ stanza/resources/prepare_resources.py
